--- src/backend/TrafficCourts/Common/Features/Mail/Templates/IDisputantUpdateRequestAcceptedTemplate.cs
@@ -1,7 +1,0 @@
-﻿using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
-
-namespace TrafficCourts.Common.Features.Mail.Templates;
-
-public interface IDisputeUpdateRequestAcceptedTemplate : IEmailTemplate<Dispute>
-{
-}
