+++ src/backend/TrafficCourts/Common/Features/Mail/Templates/IDisputeUpdateRequestAcceptedTemplate.cs
@@ -1,0 +1,7 @@
+﻿using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
+
+namespace TrafficCourts.Common.Features.Mail.Templates;
+
+public interface IDisputeUpdateRequestAcceptedTemplate : IEmailTemplate<Dispute>
+{
+}
